+++ graddog/variable.py
@@ -1,77 +1,86 @@
-# import numbers
-# from graddog.trace import Trace
-# from graddog.functions import VectorFunction
-# import numpy as np
+import numbers
+from graddog.trace import Trace
+from graddog.functions import VectorFunction
+import numpy as np
 
-# # TODO: change assert statements to ValueError exception handlers
 
-# # TODO: better name for trace function
-# # reminder: the trace function exists to create a trace object based on a function
 
-# class Variable(Trace):
+# TODO: change assert statements to ValueError exception handlers
 
-# 	def __init__(self, name, val):
-# 		# by default, the derivative of a variable with respect to itself is 1.0
+# TODO: better name for trace function
+# reminder: the trace function exists to create a trace object based on a function
 
-# 		super().__init__(name, val, {name : 1.0}, is_var = True)
-# 		self._name = name
+class Variable(Trace):
 
-# 	@property
-# 	def val(self):
-# 		'''
-# 		Returns non-public attribute _val
-# 		'''
-# 		return self._val
+	def __init__(self, name, val):
+		# by default, the derivative of a variable with respect to itself is 1.0
 
-# 	@val.setter
-# 	def val(self, new_val):
-# 		'''
-# 		This resets the _val of a Variable instance
-# 		'''
-# 		if isinstance(new_val, numbers.Number):
-# 			self._val = new_val
+		super().__init__(name, val, {name : 1.0}, is_var = True)
+		self._name = name
 
-# 		else:
-# 			raise TypeError('Value should be numerical')
+	@property
+	def val(self):
+		'''
+		Returns non-public attribute _val
+		'''
+		return self._val
 
-# def function_to_Trace(f, seed):
-# 	M = len(seed)
-# 	new_vars = list(get_vars([f'x{m+1}' for m in range(M)], seed))
-# 	new_var_names = list(map(lambda v : v._formula, new_vars))
-# 	result = f(new_vars)
+	@val.setter
+	def val(self, new_val):
+		'''
+		This resets the _val of a Variable instance
+		'''
+		if isinstance(new_val, numbers.Number):
+			self._val = new_val
 
-# 	def add_zeros_to_der(t):
-# 		for x in new_var_names:
-# 			if x not in t._der:
-# 				t._der[x] = 0.0
-# 	try:
-# 		for t in result:
-# 			add_zeros_to_der(t)
-# 		result = VectorFunction(result)
-# 	except TypeError:
-# 		add_zeros_to_der(result)
-# 	return result
+		else:
+			raise TypeError('Value should be numerical')
 
-# def get_x(seed):
+def function_to_Trace(f, seed):
+	M = len(seed)
+	new_vars = list(get_vars([f'x{m+1}' for m in range(M)], seed))
+	new_var_names = list(map(lambda v : v._formula, new_vars))
+	result = f(new_vars)
+
+	def add_zeros_to_der(t):
+		for x in new_var_names:
+			if x not in t._der:
+				t._der[x] = 0.0
+	try:
+		for t in result:
+			add_zeros_to_der(t)
+		result = VectorFunction(result)
+	except TypeError:
+		add_zeros_to_der(result)
+	return result
+
+def get_x(seed):
 # 	assert len(seed) == 1
-# 	return get_vars(['x'], seed)[0]
+	if len(seed) != 1:
+		raise ValueError('Length of seed must be 1.')
+	return get_vars(['x'], seed)[0]
 
-# def get_y(seed):
-# 	assert len(seed) == 1
-# 	return get_vars(['y'], seed)[0]
+def get_y(seed):
+	if len(seed) != 1:
+		raise ValueError('Length of seed must be 1.')
+	return get_vars(['y'], seed)[0]
 
-# def get_xy(seed):
-# 	assert len(seed) == 2
-# 	return get_vars(['x', 'y'], seed)
+def get_xy(seed):
+	if len(seed) != 2:
+		raise ValueError('Length of seed must be 2.')
+	return get_vars(['x', 'y'], seed)
 
-# def get_xyz(seed):
-# 	assert len(seed) == 3
-# 	return get_vars(['x', 'y', 'z'], seed)
+def get_xyz(seed):
+	if len(seed) != 3:
+		raise ValueError('Length of seed must be 3.')
+	return get_vars(['x', 'y', 'z'], seed)
 
-# def get_abc(seed):
-# 	assert len(seed) == 3
-# 	return get_vars(['a', 'b', 'c'], seed)
+def get_abc(seed):
+	if len(seed) != 3:
+		raise ValueError('Length of seed must be 3.')
+	return get_vars(['a', 'b', 'c'], seed)
 
-# def get_vars(names, seed):
-# 	assert len(names) == len(seed)
-# 	return list(Variable(names[i], seed[i]) for i in range(len(names)))
+def get_vars(names, seed):
+	if len(names) != len(seed):
+		raise ValueError("Lengths of seed and name lists must match")
+	return list(Variable(names[i], seed[i]) for i in range(len(names)))
