+++ tests/test_function.py
@@ -1,8 +1,8 @@
 import pytest
 import numpy as np
-from graddog.trace import Trace, get_x, get_xy, get_xyz, function_to_Trace, get_vars
+from graddog.trace import Trace
 from graddog.functions import sin, cos, tan, exp, log
-#from graddog.variable import get_x, get_xy, get_xyz, function_to_Trace, get_vars
+from graddog.variable import get_x, get_xy, get_xyz, function_to_Trace, get_vars
