+++ tests/test_variable.py
@@ -1,74 +1,104 @@
 import pytest
 import numpy as np
-# from graddog.variable import Variable
+from graddog.variable import Variable, get_x, get_xy, get_xyz, function_to_Trace, get_vars
 
-# def test_name_reset():
+
+def test_name_reset():
+    x1 = Variable('x', 2)
+    x1.name = 'new_x'
+    assert str(x1) == str(Variable('new_x', 2))
+
+def test_val_reset():
+    x1 = Variable('x', 2)
+    x1._val = 4
+    assert str(x1) == str(Variable('x', 4))
+    
+def test_getvars():
+    a, b, c = get_vars(['a', 'b', 'c'], [3, 2, 1])
+    assert a._val == 3
+    assert b._val == 2
+    assert c._val == 1
+    assert a._formula == 'a'
+    assert b._formula == 'b'
+    assert c._formula == 'c'
+    
+def test_string_input():
+    with pytest.raises(TypeError):
+        x_str = get_x(['CS107'])
+        
+def test_getx():
+    x = get_x([1])
+    assert x._val == 1
+    assert x._formula == 'x'
+   
+    
+def test_ValueError_gets():
+    with pytest.raises(ValueError):
+        len1list = [1]
+        len2list = [3,4]
+        len3list = [5,6,7]
+        x_str = get_x(len3list)
+        x, y, z = get_xyz(len2list)
+        x, y = get_xy(len1list)
+        a, b, c, d, e = get_vars(['a', 'b', 'c', 'd', 'e'], len2list)
+
+def test_val_reset_error():
+    x1 = Variable('x', 2)
+    with pytest.raises(TypeError):
+        x1._val = 'four'
+
+# def test_variable_add():
 #     x1 = Variable('x', 2)
-#     x1.name = 'new_x'
-#     assert str(x1) == str(Variable('new_x', 2))
+#     x2 = Variable('x', 2)
+#     assert str(x1 + x2) == str(Variable('x+x', 4, 2))
 
-# def test_val_reset():
+# def test_variable_radd():
+#     x1 = 4
+#     x2 = Variable('x2', 2)
+#     assert str(x1+x2) == str(Variable('x2+4', 6))
+
+# def test_variable_sub():
 #     x1 = Variable('x', 2)
-#     x1._val = 4
-#     assert str(x1) == str(Variable('x', 4))
+#     x2 = Variable('x', 2)
+#     assert str(x1-x2) == str(Variable('x-x', 0, 0))
 
-# # def test_val_reset_error():
-# #     x1 = Variable('x', 2)
-# #     with pytest.raises(TypeError):
-# #         x1._val = 'four'
+# def test_variable_sub_error():
+#     x1 = Variable('x', 4)
+#     x2 = 4
+#     assert str(x1-4) == str(Variable('x-4', 0))
 
-# # def test_variable_add():
-# #     x1 = Variable('x', 2)
-# #     x2 = Variable('x', 2)
-# #     assert str(x1 + x2) == str(Variable('x+x', 4, 2))
-
-# # def test_variable_radd():
-# #     x1 = 4
-# #     x2 = Variable('x2', 2)
-# #     assert str(x1+x2) == str(Variable('x2+4', 6))
-
-# # def test_variable_sub():
-# #     x1 = Variable('x', 2)
-# #     x2 = Variable('x', 2)
-# #     assert str(x1-x2) == str(Variable('x-x', 0, 0))
-
-# # def test_variable_sub_error():
-# #     x1 = Variable('x', 4)
-# #     x2 = 4
-# #     assert str(x1-4) == str(Variable('x-4', 0))
-
-# # def test_variable_rsub():
-# #     x1 = 4
-# #     x2 = Variable('x', 4)
-# #     assert str(x1-x2) == str(Variable('-x+4', 0, -1))
+# def test_variable_rsub():
+#     x1 = 4
+#     x2 = Variable('x', 4)
+#     assert str(x1-x2) == str(Variable('-x+4', 0, -1))
 
 
-# # def test_variable_mul():
-# #     x1 = Variable('x', 2)
-# #     x2 = Variable('x', 2)
-# #     assert str(x1 * x2) == str(Variable('x*x', 4, 4))
+# def test_variable_mul():
+#     x1 = Variable('x', 2)
+#     x2 = Variable('x', 2)
+#     assert str(x1 * x2) == str(Variable('x*x', 4, 4))
 
-# # def test_variable_div():
-# #     x1 = Variable('x', 2)
-# #     x2 = Variable('x', 2)
-# #     assert str(x1 / x2) == str(Variable('x/x', 1.0, 0))
+# def test_variable_div():
+#     x1 = Variable('x', 2)
+#     x2 = Variable('x', 2)
+#     assert str(x1 / x2) == str(Variable('x/x', 1.0, 0))
 
-# # def test_variable_div_error():
-# #     x1 = Variable('x', 3)
-# #     x2 = 3
-# #     assert str(x1/x2) == str(Variable('x/3', 1, (1/3)))
+# def test_variable_div_error():
+#     x1 = Variable('x', 3)
+#     x2 = 3
+#     assert str(x1/x2) == str(Variable('x/3', 1, (1/3)))
 
-# # def test_variable_rdiv():
-# #     x1 = 3
-# #     x2 = Variable('x', 3)
-# #     assert str(x1/x2) == str(Variable('3/x', 1, (-1/3)))
+# def test_variable_rdiv():
+#     x1 = 3
+#     x2 = Variable('x', 3)
+#     assert str(x1/x2) == str(Variable('3/x', 1, (-1/3)))
 
-# # def test_variable_neg():
-# #     x1 = Variable('x', 2)
-# #     x2 = -x1
-# #     assert "-" + x1._formula ==  x2._formula 
-# #     assert -x1._val == x2._val
-# #     assert -x1._der == x2._der
+# def test_variable_neg():
+#     x1 = Variable('x', 2)
+#     x2 = -x1
+#     assert "-" + x1._formula ==  x2._formula 
+#     assert -x1._val == x2._val
+#     assert -x1._der == x2._der
 
 # def test_variable_pow():
 #     x1 = Variable('x', 2)
@@ -91,6 +121,6 @@ import numpy as np
 #     f = 3*x1**2 + 2*x1 + 5
 #     assert f._der['x'] == 32
 
-# # def test_string_input():
-# #     with pytest.raises(TypeError):
-# #         x_str = Variable('s', 'hello World')
+# def test_string_input():
+#     with pytest.raises(TypeError):
+#         x_str = Variable('s', 'hello World')
